requirements.txt
absl-py==0.3.0
aiml==0.9.1
appdirs==1.4.3
astor==0.7.1
btrfsutil==1.0.0
CacheControl==0.12.5
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
distlib==0.2.7
distro==1.3.0
dj-database-url==0.5.0
Django==2.0.5
django-appconf==1.0.2
django-compressor==2.2
django-tinymce==2.7.0
django-tinymce4-lite==1.7.2
django-widget-tweaks==1.4.2
doco==0.0.4
editdistance==0.4
future==0.16.0
gast==0.2.0
grpcio==1.13.0
Guake==3.3.2
gunicorn==19.8.1
html5lib==1.0.1
idna==2.6
Janome==0.3.6
jsmin==2.2.2
lensfun==0.3.2
line-bot-sdk==1.5.0
lockfile==0.12.2
Markdown==2.6.11
mecab-python3==0.7
msgpack==0.5.6
numpy==1.14.3
packaging==17.1
pandas==0.22.0
pbr==4.1.0
Pillow==5.2.0
progress==1.4
protobuf==3.6.0
psycopg2==2.7.4
psycopg2-binary==2.7.5
pycairo==1.17.0
pygobject==3.28.3
pyparsing==2.2.0
python-aiml==0.9.1
python-dateutil==2.7.2
pytoml==0.1.16
pytz==2018.4
rcssmin==1.0.6
requests==2.18.4
retrying==1.3.3
rjsmin==1.0.12
simplejson==3.14.0
six==1.11.0
team==1.0
tensorboard==1.9.0
tensorflow==1.9.0
termcolor==1.1.0
urllib3==1.22
webencodings==0.5.1
Werkzeug==0.14.1
whitenoise==3.3.1
